+++ src/main/java/com/smunity/api/domain/petition/service/impl/AgreementServiceImpl.java
@@ -9,7 +9,7 @@ import com.smunity.api.domain.petition.repository.AgreementRepository;
 import com.smunity.api.domain.petition.repository.PetitionRepository;
 import com.smunity.api.domain.petition.service.AgreementService;
 import com.smunity.api.global.config.security.JwtTokenProvider;
-import com.smunity.api.global.exception.CustomException;
+import com.smunity.api.global.error.exception.RestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -26,7 +26,7 @@ public class AgreementServiceImpl implements AgreementService {
 
     @Override
     public List<AgreementDto> getAgreementUsersByPetitionId(Long petitionId) {
-        petitionRepository.findById(petitionId).orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+        petitionRepository.findById(petitionId).orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         List<Agreement> agreementList = agreementRepository.findAllByPetitionId(petitionId);
         for (Agreement agreement: agreementList)
             System.out.println(agreement);
@@ -36,14 +36,14 @@ public class AgreementServiceImpl implements AgreementService {
     @Override
     public AgreementDto createAgreement(Long petitionId, String token) {
         Petition petition = petitionRepository.findById(petitionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (jwtTokenProvider.getUsername(token).equals(petition.getAuthor().getUsername()))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         User user = userRepository.getByUsername(jwtTokenProvider.getUsername(token));
         if (agreementRepository.existsByPetitionIdAndUserId(petitionId, user.getId()))
-            throw new CustomException(HttpStatus.CONFLICT);
+            throw new RestException(HttpStatus.CONFLICT);
         Agreement agreement = AgreementDto.toEntity(user, petition);
         Agreement saveAgreement = agreementRepository.save(agreement);
         return AgreementDto.toDto(saveAgreement);
