+++ src/main/java/com/smunity/api/domain/petition/service/impl/RespondServiceImpl.java
@@ -9,7 +9,7 @@ import com.smunity.api.domain.petition.repository.RespondRepository;
 import com.smunity.api.domain.petition.repository.PetitionRepository;
 import com.smunity.api.domain.petition.service.RespondService;
 import com.smunity.api.global.config.security.JwtTokenProvider;
-import com.smunity.api.global.exception.CustomException;
+import com.smunity.api.global.error.exception.RestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -26,13 +26,13 @@ public class RespondServiceImpl implements RespondService {
     @Override
     public RespondDto createRespond(Long petitionId, RespondDto respondDto, String token) {
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         if (respondRepository.existsByPetitionId(petitionId))
-            throw new CustomException(HttpStatus.CONFLICT);
+            throw new RestException(HttpStatus.CONFLICT);
         Petition petition = petitionRepository.findById(petitionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         String username = jwtTokenProvider.getUsername(token);
         User user = userRepository.getByUsername(username);
         Respond respond = respondDto.toEntity(user, petition);
@@ -43,18 +43,18 @@ public class RespondServiceImpl implements RespondService {
     @Override
     public RespondDto getRespondByPetitionId(Long petitionId) {
         Respond respond = respondRepository.findByPetitionId(petitionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         return RespondDto.toDto(respond);
     }
 
     @Override
     public RespondDto updateRespond(Long petitionId, RespondDto respondDto, String token) {
         Respond respond = respondRepository.findByPetitionId(petitionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         respond.setContent(respondDto.getContent());
         Respond changedRespond = respondRepository.save(respond);
         return RespondDto.toDto(changedRespond);
@@ -63,11 +63,11 @@ public class RespondServiceImpl implements RespondService {
     @Override
     public void deleteRespond(Long petitionId, String token) {
         Respond respond = respondRepository.findByPetitionId(petitionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NO_CONTENT));
+                .orElseThrow(() -> new RestException(HttpStatus.NO_CONTENT));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         respondRepository.delete(respond);
     }
 }
