+++ src/main/java/com/smunity/api/domain/question/service/impl/AnswerServiceImpl.java
@@ -9,7 +9,7 @@ import com.smunity.api.domain.question.repository.AnswerRepository;
 import com.smunity.api.domain.question.repository.QuestionRepository;
 import com.smunity.api.domain.question.service.AnswerService;
 import com.smunity.api.global.config.security.JwtTokenProvider;
-import com.smunity.api.global.exception.CustomException;
+import com.smunity.api.global.error.exception.RestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -26,13 +26,13 @@ public class AnswerServiceImpl implements AnswerService {
     @Override
     public AnswerDto createAnswer(Long questionId, AnswerDto answerDto, String token) {
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         if (answerRepository.existsByQuestionId(questionId))
-            throw new CustomException(HttpStatus.CONFLICT);
+            throw new RestException(HttpStatus.CONFLICT);
         Question question = questionRepository.findById(questionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         String username = jwtTokenProvider.getUsername(token);
         User user = userRepository.getByUsername(username);
         Answer answer = answerDto.toEntity(user, question);
@@ -44,18 +44,18 @@ public class AnswerServiceImpl implements AnswerService {
     @Override
     public AnswerDto getAnswerQuestionId(Long questionId) {
         Answer answer = answerRepository.findByQuestionId(questionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         return AnswerDto.toDto(answer);
     }
 
     @Override
     public AnswerDto updateAnswer(Long questionId, AnswerDto answerDto, String token) {
         Answer answer = answerRepository.findByQuestionId(questionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         answer.setContent(answerDto.getContent());
         Answer changedAnswer = answerRepository.save(answer);
         return AnswerDto.toDto(changedAnswer);
@@ -64,11 +64,11 @@ public class AnswerServiceImpl implements AnswerService {
     @Override
     public void deleteAnswer(Long questionId, String token) {
         Answer answer = answerRepository.findByQuestionId(questionId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NO_CONTENT));
+                .orElseThrow(() -> new RestException(HttpStatus.NO_CONTENT));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         answerRepository.delete(answer);
     }
 }
