+++ src/main/java/com/smunity/api/domain/question/service/impl/QuestionServiceImpl.java
@@ -7,7 +7,7 @@ import com.smunity.api.domain.question.dto.QuestionDto;
 import com.smunity.api.domain.question.repository.QuestionRepository;
 import com.smunity.api.domain.question.service.QuestionService;
 import com.smunity.api.global.config.security.JwtTokenProvider;
-import com.smunity.api.global.exception.CustomException;
+import com.smunity.api.global.error.exception.RestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -30,7 +30,7 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     public QuestionDto createQuestion(QuestionDto questionDto, String token) {
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         String username = jwtTokenProvider.getUsername(token);
         User user = userRepository.getByUsername(username);
         Question question = questionDto.toEntity(user);
@@ -42,18 +42,18 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     public QuestionDto getQuestionById(Long id) {
         Question question = questionRepository.findById(id)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         return QuestionDto.toDto(question);
     }
 
     @Override
     public QuestionDto updateQuestion(Long id, QuestionDto questionDto, String token) {
         Question question = questionRepository.findById(id)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND));
+                .orElseThrow(() -> new RestException(HttpStatus.NOT_FOUND));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getUsername(token).equals(question.getAuthor().getUsername()) && !jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         question.setSubject(questionDto.getSubject());
         question.setContent(questionDto.getContent());
         question.setAnonymous(questionDto.getAnonymous());
@@ -64,11 +64,11 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     public void deleteQuestion(Long id, String token) {
         Question question = questionRepository.findById(id)
-                .orElseThrow(() -> new CustomException(HttpStatus.NO_CONTENT));
+                .orElseThrow(() -> new RestException(HttpStatus.NO_CONTENT));
         if (!jwtTokenProvider.validateToken(token))
-            throw new CustomException(HttpStatus.UNAUTHORIZED);
+            throw new RestException(HttpStatus.UNAUTHORIZED);
         if (!jwtTokenProvider.getUsername(token).equals(question.getAuthor().getUsername()) && !jwtTokenProvider.getIsSuperuser(token))
-            throw new CustomException(HttpStatus.FORBIDDEN);
+            throw new RestException(HttpStatus.FORBIDDEN);
         questionRepository.delete(question);
     }
 }
