+++ src/main/java/com/smunity/api/global/config/security/JwtAuthenticationFilter.java
@@ -1,5 +1,6 @@
 package com.smunity.api.global.config.security;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -10,12 +11,9 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 
+@RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
     private final JwtTokenProvider jwtTokenProvider;
-
-    public JwtAuthenticationFilter(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest servletRequest,
