+++ src/main/java/com/smunity/api/global/error/exception/RestException.java
@@ -1,4 +1,4 @@
-package com.smunity.api.global.exception;
+package com.smunity.api.global.error.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import org.springframework.http.HttpStatus;
 
 @AllArgsConstructor
 @Getter
-public class CustomException extends RuntimeException {
+public class RestException extends RuntimeException {
     private HttpStatus httpStatus;
 }
